requirements.txt
-e .
PyYAML